able to restrict z3 caller memory usage

+++ Domain/src/main/scala/peal/domain/Pol.scala
@@ -4,7 +4,7 @@ import scala.collection.JavaConversions._
 import peal.domain.operator.Operators
 
 case class Pol(rules: java.util.List[Rule], operator: Operators, defaultScore: BigDecimal, name: String = "") extends PolicySet {
-  override def toString: String = operator + " (" + rules.mkString(" ") + ") default " + "%.2f".format(defaultScore)
+  override def toString: String = operator + " (" + rules.mkString(" ") + ") default " + "%.2f".format(defaultScore.doubleValue())
 
   //needed by Java classes
   def getName = name
